pyproject.toml
[project]
name = "streamlit-ad-app"
version = "0.1.0"
description = "AI anomaly detection program generator using Streamlit"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.45.0,<2.0.0",
    "anthropic>=0.40.0",
    "torch>=2.5.0,<3.0.0",
    "torchvision>=0.20.0,<1.0.0",
    "transformers>=4.46.0,<5.0.0",
    "pillow>=10.2.0",
    "numpy>=1.26.0",
    "psutil>=5.9.0",
    "fastapi>=0.115.0,<1.0.0",
    "uvicorn>=0.34.0,<1.0.0",
    "pandas>=2.2.0,<3.0.0",
    "matplotlib>=3.10.0,<4.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "black>=25.9.0",
    "flake8>=7.3.0",
    "isort>=6.0.1",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
]
